requirements.txt
django==5.0.4
djangorestframework==3.15.2
django-filter
python-dotenv==1.0.1
